# Optimized topk for topk=1 (Llama-4)