@njhill 2025-04-11
@mgoin could we use this for other moes too? e.g. in https://github.com/vllm-project/vllm/blob/main/vllm/model_executor/layers/fused_moe/fused_moe.py#L886?