@mgoin 2025-04-11
@njhill unfortunately most other moes do not use a topk=1 AFAIK, but maybe the overhead is minimal enough to use just in case